+++ src/VoltageSource.java
@@ -9,8 +9,8 @@ public class VoltageSource extends Element {
         node1 = info[1];
         node2 = info[2];
         if (info[3].charAt(info[3].length() - 1) >= '0' && info[3].charAt(info[3].length() - 1) <= '9')
-            dc = Integer.parseInt(info[3]);
-        else dc = Integer.parseInt(info[3].substring(0, info[3].length() - 1));
+            dc = Double.parseDouble(info[3]);
+        else dc = Double.parseDouble(info[3].substring(0, info[3].length() - 1));
         switch (info[3].charAt(info[3].length() - 1)) {
             case 'p':
                 dc *= 1e-12;
@@ -35,8 +35,8 @@ public class VoltageSource extends Element {
                 break;
         }
         if (info[4].charAt(info[4].length() - 1) >= '0' && info[4].charAt(info[4].length() - 1) <= '9')
-            ac = Integer.parseInt(info[4]);
-        else ac = Integer.parseInt(info[4].substring(0, info[4].length() - 1));
+            ac = Double.parseDouble(info[4]);
+        else ac = Double.parseDouble(info[4].substring(0, info[4].length() - 1));
         switch (info[4].charAt(info[4].length() - 1)) {
             case 'p':
                 ac *= 1e-12;
@@ -60,9 +60,7 @@ public class VoltageSource extends Element {
                 ac *= 1e9;
                 break;
         }
-        frequncey = Integer.parseInt(info[5]);
-        phase = Integer.parseInt(info[6]);
-
-
+        frequncey = Double.parseDouble(info[5]);
+        phase = Double.parseDouble(info[6]);
     }
 }
